Refactor bot to use Baileys for WhatsApp connection and update README
