+++ README.md
@@ -1,16 +1,21 @@
 # bot.whatssap
-Criei um bot para whatssap para baixar videos do youtuber quando mandar link para ele.
-<br>
-No momento utilizei as api whatsapp-web.js para linkar a conta do whatssap é o ytdl-core para baixar o video no 
-equipamento e depois enviar para o usuario do outro lado!
-Até o momento ele está bem simples mais funcionando , caso queira usar instale as dependencias é inicie o programa link com o whatssap e pronto!
-<br>
-Caso tenha gostado ou queira dar uma dica de melhoria pode ficar a vontade! ; )
 
-proxima versão utilizar o Baileys para gerenciamento de conta zap.
-https://github.com/WhiskeySockets/Baileys
+Criei um bot para WhatsApp para baixar vídeos do YouTube quando mandar um link para ele.
 
-Requisitos:
-            baixar videos do youtuber
-            baixar musicas do youtuber
-            
+No momento, utilizei as APIs `whatsapp-web.js` para linkar a conta do WhatsApp e `ytdl-core` para baixar o vídeo no equipamento e depois enviar para o usuário do outro lado. Até o momento, ele está bem simples, mas funcionando. Caso queira usar, instale as dependências e inicie o programa, link com o WhatsApp e pronto!
+
+Caso tenha gostado ou queira dar uma dica de melhoria, pode ficar à vontade! 😉
+
+## Próxima Versão
+
+Na próxima versão, utilizarei o Baileys para gerenciamento de conta no WhatsApp.
+[Baileys GitHub](https://github.com/WhiskeySockets/Baileys)
+
+## Requisitos
+
+- Baixar vídeos do YouTube
+- Baixar músicas do YouTube
+
+## Como Usar
+
+1. Clone o repositório:
