+++ package.json
@@ -5,20 +5,20 @@
   "main": "CodigoPrincipal.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon index.js"
+    "dev": "nodemon --exec ts-node index.ts"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@hapi/boom": "^10.0.1",
     "@whiskeysockets/baileys": "^6.7.9",
-    "baileys": "^6.7.9",
     "long": "^5.2.3",
-    "nodemon": "^3.1.7",
     "qrcode-terminal": "^0.12.0"
   },
   "devDependencies": {
-    "typescript": "^5.6.3"
+    "nodemon": "^3.1.7",
+    "typescript": "^5.6.3",
+    "ts-node": "^10.9.1"
   },
   "keywords": []
 }
